+++ examples/M1/M1_export.py
@@ -3,8 +3,8 @@ from netpyne import sim  # import netpyne sim module
 
 np = M1.netParams
 print("********************\n*\n*  Note: setting noise to 1, since noise can only be 0 or 1 in NeuroML export currently!\n*\n********************")
-np.popParams['background_E']['noise'] = 1
-np.popParams['background_I']['noise'] = 1
+np.stimSourceParams['background_E']['noise'] = 1
+np.stimSourceParams['background_I']['noise'] = 1
 
 sim.createExportNeuroML2(netParams = np, 
                        simConfig = M1.simConfig,